requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.1
pymongo==4.16.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.22